+++ src/utils/helpers.ts
@@ -4,5 +4,14 @@ import React from 'react';
 export function stopPropagation<ElementType = unknown>(
   e: React.MouseEvent<ElementType, MouseEvent>,
 ): void {
+  console.log('stopping');
   e.stopPropagation();
 }
+
+/**
+ * A handy jQuery-like selector that makes accessing the DOM
+ * easier, but won't include all of the overhead of the actual
+ * jQuery library itself.
+ */
+export const $: typeof document.querySelector =
+  document.querySelector.bind(document);
